PRSD-404: Make validation service generic
