+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/dataModels/ValidatableDataModel.kt
@@ -1,0 +1,9 @@
+package uk.gov.communities.prsdb.webapp.models.dataModels
+
+interface ValidatableDataModel {
+    val errorFieldMap: Map<String, String>
+        get() = mapOf()
+
+    val errorPrecedenceList: List<String>
+        get() = listOf()
+}
